adding the favicon to the starter and deactivating the session (from cdiconversation)
