+++ component-starter-server/src/main/frontend/config/webpack.config.dev.js
@@ -32,280 +32,281 @@ const env = getClientEnvironment(publicUrl);
 // It is focused on developer experience and fast rebuilds.
 // The production configuration is different and lives in a separate file.
 module.exports = {
-	// You may want 'eval' instead if you prefer to see the compiled output in DevTools.
-	// See the discussion in https://github.com/facebookincubator/create-react-app/issues/343.
-	devtool: 'cheap-module-source-map',
-	// These are the "entry points" to our application.
-	// This means they will be the "root" imports that are included in JS bundle.
-	// The first two entry points enable "hot" CSS and auto-refreshes for JS.
-	entry: [
-		// Include an alternative client for WebpackDevServer. A client's job is to
-		// connect to WebpackDevServer by a socket and get notified about changes.
-		// When you save a file, the client will either apply hot updates (in case
-		// of CSS changes), or refresh the page (in case of JS changes). When you
-		// make a syntax error, this client will display a syntax error overlay.
-		// Note: instead of the default WebpackDevServer client, we use a custom one
-		// to bring better experience for Create React App users. You can replace
-		// the line below with these two lines if you prefer the stock client:
-		// require.resolve('webpack-dev-server/client') + '?/',
-		// require.resolve('webpack/hot/dev-server'),
-		require.resolve('react-dev-utils/webpackHotDevClient'),
-		// We ship a few polyfills by default:
-		require.resolve('./polyfills'),
-		// Finally, this is your app's code:
-		paths.appIndexJs,
-		// We include the app code last so that if there is a runtime error during
-		// initialization, it doesn't blow up the WebpackDevServer client, and
-		// changing JS code would still trigger a refresh.
-	],
-	output: {
-		// Next line is not used in dev but WebpackDevServer crashes without it:
-		path: paths.appBuild,
-		// Add /* filename */ comments to generated require()s in the output.
-		pathinfo: true,
-		// This does not produce a real file. It's just the virtual path that is
-		// served by WebpackDevServer in development. This is the JS bundle
-		// containing code from all our entry points, and the Webpack runtime.
-		filename: 'static/js/bundle.js',
-		// There are also additional JS chunk files if you use code splitting.
-		chunkFilename: 'static/js/[name].chunk.js',
-		// This is the URL that app is served from. We use "/" in development.
-		publicPath: publicPath,
-		// Point sourcemap entries to original disk location (format as URL on Windows)
-		devtoolModuleFilenameTemplate: info =>
-			path.resolve(info.absoluteResourcePath).replace(/\\/g, '/'),
-	},
-	resolve: {
-		// This allows you to set a fallback for where Webpack should look for modules.
-		// We placed these paths second because we want `node_modules` to "win"
-		// if there are any conflicts. This matches Node resolution mechanism.
-		// https://github.com/facebookincubator/create-react-app/issues/253
-		modules: ['node_modules', paths.appNodeModules].concat(
-			// It is guaranteed to exist because we tweak it in `env.js`
-			process.env.NODE_PATH.split(path.delimiter).filter(Boolean)
-		),
-		// These are the reasonable defaults supported by the Node ecosystem.
-		// We also include JSX as a common component filename extension to support
-		// some tools, although we do not recommend using it, see:
-		// https://github.com/facebookincubator/create-react-app/issues/290
-		// `web` extension prefixes have been added for better support
-		// for React Native Web.
-		extensions: ['.web.js', '.js', '.json', '.web.jsx', '.jsx'],
-		alias: {
+  // You may want 'eval' instead if you prefer to see the compiled output in DevTools.
+  // See the discussion in https://github.com/facebookincubator/create-react-app/issues/343.
+  devtool: 'cheap-module-source-map',
+  // These are the "entry points" to our application.
+  // This means they will be the "root" imports that are included in JS bundle.
+  // The first two entry points enable "hot" CSS and auto-refreshes for JS.
+  entry: [
+    // Include an alternative client for WebpackDevServer. A client's job is to
+    // connect to WebpackDevServer by a socket and get notified about changes.
+    // When you save a file, the client will either apply hot updates (in case
+    // of CSS changes), or refresh the page (in case of JS changes). When you
+    // make a syntax error, this client will display a syntax error overlay.
+    // Note: instead of the default WebpackDevServer client, we use a custom one
+    // to bring better experience for Create React App users. You can replace
+    // the line below with these two lines if you prefer the stock client:
+    // require.resolve('webpack-dev-server/client') + '?/',
+    // require.resolve('webpack/hot/dev-server'),
+    require.resolve('react-dev-utils/webpackHotDevClient'),
+    // We ship a few polyfills by default:
+    require.resolve('./polyfills'),
+    // Finally, this is your app's code:
+    paths.appIndexJs,
+    // We include the app code last so that if there is a runtime error during
+    // initialization, it doesn't blow up the WebpackDevServer client, and
+    // changing JS code would still trigger a refresh.
+  ],
+  output: {
+    // Next line is not used in dev but WebpackDevServer crashes without it:
+    path: paths.appBuild,
+    // Add /* filename */ comments to generated require()s in the output.
+    pathinfo: true,
+    // This does not produce a real file. It's just the virtual path that is
+    // served by WebpackDevServer in development. This is the JS bundle
+    // containing code from all our entry points, and the Webpack runtime.
+    filename: 'static/js/bundle.js',
+    // There are also additional JS chunk files if you use code splitting.
+    chunkFilename: 'static/js/[name].chunk.js',
+    // This is the URL that app is served from. We use "/" in development.
+    publicPath: publicPath,
+    // Point sourcemap entries to original disk location (format as URL on Windows)
+    devtoolModuleFilenameTemplate: info =>
+      path.resolve(info.absoluteResourcePath).replace(/\\/g, '/'),
+  },
+  resolve: {
+    // This allows you to set a fallback for where Webpack should look for modules.
+    // We placed these paths second because we want `node_modules` to "win"
+    // if there are any conflicts. This matches Node resolution mechanism.
+    // https://github.com/facebookincubator/create-react-app/issues/253
+    modules: ['node_modules', paths.appNodeModules].concat(
+      // It is guaranteed to exist because we tweak it in `env.js`
+      process.env.NODE_PATH.split(path.delimiter).filter(Boolean)
+    ),
+    // These are the reasonable defaults supported by the Node ecosystem.
+    // We also include JSX as a common component filename extension to support
+    // some tools, although we do not recommend using it, see:
+    // https://github.com/facebookincubator/create-react-app/issues/290
+    // `web` extension prefixes have been added for better support
+    // for React Native Web.
+    extensions: ['.web.js', '.js', '.json', '.web.jsx', '.jsx'],
+    alias: {
 
-			// Support React Native Web
-			// https://www.smashingmagazine.com/2016/08/a-glimpse-into-the-future-with-react-native-for-web/
-			'react-native': 'react-native-web',
-		},
-		plugins: [
-			// Prevents users from importing files from outside of src/ (or node_modules/).
-			// This often causes confusion because we only process files within src/ with babel.
-			// To fix this, we prevent you from importing files out of src/ -- if you'd like to,
-			// please link the files into your node_modules/ and let module-resolution kick in.
-			// Make sure your source files are compiled, as they will not be processed in any way.
-			new ModuleScopePlugin(paths.appSrc, [paths.appPackageJson]),
-		],
-	},
-	module: {
-		strictExportPresence: true,
-		rules: [
-			// TODO: Disable require.ensure as it's not a standard language feature.
-			// We are waiting for https://github.com/facebookincubator/create-react-app/issues/2176.
-			// { parser: { requireEnsure: false } },
+      // Support React Native Web
+      // https://www.smashingmagazine.com/2016/08/a-glimpse-into-the-future-with-react-native-for-web/
+      'react-native': 'react-native-web',
+    },
+    plugins: [
+      // Prevents users from importing files from outside of src/ (or node_modules/).
+      // This often causes confusion because we only process files within src/ with babel.
+      // To fix this, we prevent you from importing files out of src/ -- if you'd like to,
+      // please link the files into your node_modules/ and let module-resolution kick in.
+      // Make sure your source files are compiled, as they will not be processed in any way.
+      new ModuleScopePlugin(paths.appSrc, [paths.appPackageJson]),
+    ],
+  },
+  module: {
+    strictExportPresence: true,
+    rules: [
+      // TODO: Disable require.ensure as it's not a standard language feature.
+      // We are waiting for https://github.com/facebookincubator/create-react-app/issues/2176.
+      // { parser: { requireEnsure: false } },
 
-			// First, run the linter.
-			// It's important to do this before Babel processes the JS.
-			{
-				test: /\.(js|jsx)$/,
-				enforce: 'pre',
-				use: [
-					{
-						options: {
-							formatter: eslintFormatter,
-							eslintPath: require.resolve('eslint'),
+      // First, run the linter.
+      // It's important to do this before Babel processes the JS.
+      {
+        test: /\.(js|jsx)$/,
+        enforce: 'pre',
+        use: [
+          {
+            options: {
+              formatter: eslintFormatter,
+              eslintPath: require.resolve('eslint'),
 
-						},
-						loader: require.resolve('eslint-loader'),
-					},
-				],
-				include: paths.appSrc,
-			},
-			{
-				// "oneOf" will traverse all following loaders until one will
-				// match the requirements. When no loader matches it will fall
-				// back to the "file" loader at the end of the loader list.
-				oneOf: [
-					// "url" loader works like "file" loader except that it embeds assets
-					// smaller than specified limit in bytes as data URLs to avoid requests.
-					// A missing `test` is equivalent to a match.
-					{
-						test: [/\.bmp$/, /\.gif$/, /\.jpe?g$/, /\.png$/],
-						loader: require.resolve('url-loader'),
-						options: {
-							limit: 10000,
-							name: 'static/media/[name].[hash:8].[ext]',
-						},
-					},
-					// Process JS with Babel.
-					{
-						test: /\.(js|jsx)$/,
-						include: paths.appSrc,
-						loader: require.resolve('babel-loader'),
-						options: {
+            },
+            loader: require.resolve('eslint-loader'),
+          },
+        ],
+        include: paths.appSrc,
+      },
+      {
+        // "oneOf" will traverse all following loaders until one will
+        // match the requirements. When no loader matches it will fall
+        // back to the "file" loader at the end of the loader list.
+        oneOf: [
+          // "url" loader works like "file" loader except that it embeds assets
+          // smaller than specified limit in bytes as data URLs to avoid requests.
+          // A missing `test` is equivalent to a match.
+          {
+            test: [/\.bmp$/, /\.gif$/, /\.jpe?g$/, /\.png$/],
+            loader: require.resolve('url-loader'),
+            options: {
+              limit: 10000,
+              name: 'static/media/[name].[hash:8].[ext]',
+            },
+          },
+          // Process JS with Babel.
+          {
+            test: /\.(js|jsx)$/,
+            include: paths.appSrc,
+            loader: require.resolve('babel-loader'),
+            options: {
 
-							// This is a feature of `babel-loader` for webpack (not Babel itself).
-							// It enables caching results in ./node_modules/.cache/babel-loader/
-							// directory for faster rebuilds.
-							cacheDirectory: true,
-						},
-					},
-					// "postcss" loader applies autoprefixer to our CSS.
-					// "css" loader resolves paths in CSS and adds assets as dependencies.
-					// "style" loader turns CSS into JS modules that inject <style> tags.
-					// In production, we use a plugin to extract that CSS to a file, but
-					// in development "style" loader enables hot editing of CSS.
-					{
-						test: /\.css$/,
-						use: [
-							require.resolve('style-loader'),
-							{
-								loader: require.resolve('css-loader'),
-								options: {
-									importLoaders: 1,
-								},
-							},
-							{
-								loader: require.resolve('postcss-loader'),
-								options: {
-									// Necessary for external CSS imports to work
-									// https://github.com/facebookincubator/create-react-app/issues/2677
-									ident: 'postcss',
-									plugins: () => [
-										require('postcss-flexbugs-fixes'),
-										autoprefixer({
-											browsers: [
-												'>1%',
-												'last 4 versions',
-												'Firefox ESR',
-												'not ie < 9', // React doesn't support IE8 anyway
-											],
-											flexbox: 'no-2009',
-										}),
-									],
-								},
-							},
-						],
-					},
-					{
-						test: /\.scss$/,
-						use: [
-							require.resolve('style-loader'),
-							{
-								loader: require.resolve('css-loader'),
-								options: {
-									sourceMap: true,
-									modules: true,
-									importLoaders: 1,
-									localIdentName: '[name]__[local]___[hash:base64:5]',
-								},
-							},
-							{
-								loader: require.resolve('postcss-loader'),
-								options: {
-									// Necessary for external CSS imports to work
-									// https://github.com/facebookincubator/create-react-app/issues/2677
-									ident: 'postcss',
-									plugins: () => [
-										require('postcss-flexbugs-fixes'),
-										autoprefixer({
-											browsers: [
-												'>1%',
-												'last 4 versions',
-												'Firefox ESR',
-												'not ie < 9', // React doesn't support IE8 anyway
-											],
-											flexbox: 'no-2009',
-										}),
-									],
-									sourceMap: true,
-								},
-							},
-							{ loader: 'resolve-url-loader' },
-							{
-								loader: require.resolve('sass-loader'),
-								options: { sourceMap: true, data: SASS_DATA }
-							}
-						],
-						// include: /@talend/,
-					},
-					// "file" loader makes sure those assets get served by WebpackDevServer.
-					// When you `import` an asset, you get its (virtual) filename.
-					// In production, they would get copied to the `build` folder.
-					// This loader don't uses a "test" so it will catch all modules
-					// that fall through the other loaders.
-					{
-						// Exclude `js` files to keep "css" loader working as it injects
-						// it's runtime that would otherwise processed through "file" loader.
-						// Also exclude `html` and `json` extensions so they get processed
-						// by webpacks internal loaders.
-						exclude: [/\.js$/, /\.html$/, /\.json$/],
-						loader: require.resolve('file-loader'),
-						options: {
-							name: 'static/media/[name].[hash:8].[ext]',
-						},
-					},
-				],
-			},
-			// ** STOP ** Are you adding a new loader?
-			// Make sure to add the new loader(s) before the "file" loader.
-		],
-	},
-	plugins: [
-		// Makes some environment variables available in index.html.
-		// The public URL is available as %PUBLIC_URL% in index.html, e.g.:
-		// <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
-		// In development, this will be an empty string.
-		new InterpolateHtmlPlugin(env.raw),
-		// Generates an `index.html` file with the <script> injected.
-		new HtmlWebpackPlugin({
-			inject: true,
-			template: paths.appHtml,
-		}),
-		// Add module names to factory functions so they appear in browser profiler.
-		new webpack.NamedModulesPlugin(),
-		// Makes some environment variables available to the JS code, for example:
-		// if (process.env.NODE_ENV === 'development') { ... }. See `./env.js`.
-		new webpack.DefinePlugin(env.stringified),
-		// This is necessary to emit hot updates (currently CSS only):
-		new webpack.HotModuleReplacementPlugin(),
-		// Watcher doesn't work well if you mistype casing in a path so we use
-		// a plugin that prints an error when you attempt to do this.
-		// See https://github.com/facebookincubator/create-react-app/issues/240
-		new CaseSensitivePathsPlugin(),
-		// If you require a missing module and then `npm install` it, you still have
-		// to restart the development server for Webpack to discover it. This plugin
-		// makes the discovery automatic so you don't have to restart.
-		// See https://github.com/facebookincubator/create-react-app/issues/186
-		new WatchMissingNodeModulesPlugin(paths.appNodeModules),
-		// Moment.js is an extremely popular library that bundles large locale files
-		// by default due to how Webpack interprets its code. This is a practical
-		// solution that requires the user to opt into importing specific locales.
-		// https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
-		// You can remove this if you don't use Moment.js:
-		new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
-	],
-	// Some libraries import Node modules but don't use them in the browser.
-	// Tell Webpack to provide empty mocks for them so importing them works.
-	node: {
-		dgram: 'empty',
-		fs: 'empty',
-		net: 'empty',
-		tls: 'empty',
-	},
-	// Turn off performance hints during development because we don't do any
-	// splitting or minification in interest of speed. These warnings become
-	// cumbersome.
-	performance: {
-		hints: false,
-	},
+              // This is a feature of `babel-loader` for webpack (not Babel itself).
+              // It enables caching results in ./node_modules/.cache/babel-loader/
+              // directory for faster rebuilds.
+              cacheDirectory: true,
+            },
+          },
+          // "postcss" loader applies autoprefixer to our CSS.
+          // "css" loader resolves paths in CSS and adds assets as dependencies.
+          // "style" loader turns CSS into JS modules that inject <style> tags.
+          // In production, we use a plugin to extract that CSS to a file, but
+          // in development "style" loader enables hot editing of CSS.
+          {
+            test: /\.css$/,
+            use: [
+              require.resolve('style-loader'),
+              {
+                loader: require.resolve('css-loader'),
+                options: {
+                  importLoaders: 1,
+                },
+              },
+              {
+                loader: require.resolve('postcss-loader'),
+                options: {
+                  // Necessary for external CSS imports to work
+                  // https://github.com/facebookincubator/create-react-app/issues/2677
+                  ident: 'postcss',
+                  plugins: () => [
+                    require('postcss-flexbugs-fixes'),
+                    autoprefixer({
+                      browsers: [
+                        '>1%',
+                        'last 4 versions',
+                        'Firefox ESR',
+                        'not ie < 9', // React doesn't support IE8 anyway
+                      ],
+                      flexbox: 'no-2009',
+                    }),
+                  ],
+                },
+              },
+            ],
+          },
+          {
+            test: /\.scss$/,
+            use: [
+              require.resolve('style-loader'),
+              {
+                loader: require.resolve('css-loader'),
+                options: {
+                  sourceMap: true,
+                  modules: true,
+                  importLoaders: 1,
+                  localIdentName: '[name]__[local]___[hash:base64:5]',
+                },
+              },
+              {
+                loader: require.resolve('postcss-loader'),
+                options: {
+                  // Necessary for external CSS imports to work
+                  // https://github.com/facebookincubator/create-react-app/issues/2677
+                  ident: 'postcss',
+                  plugins: () => [
+                    require('postcss-flexbugs-fixes'),
+                    autoprefixer({
+                      browsers: [
+                        '>1%',
+                        'last 4 versions',
+                        'Firefox ESR',
+                        'not ie < 9', // React doesn't support IE8 anyway
+                      ],
+                      flexbox: 'no-2009',
+                    }),
+                  ],
+                  sourceMap: true,
+                },
+              },
+              { loader: 'resolve-url-loader' },
+              {
+                loader: require.resolve('sass-loader'),
+                options: { sourceMap: true, data: SASS_DATA }
+              }
+            ],
+            // include: /@talend/,
+          },
+          // "file" loader makes sure those assets get served by WebpackDevServer.
+          // When you `import` an asset, you get its (virtual) filename.
+          // In production, they would get copied to the `build` folder.
+          // This loader don't uses a "test" so it will catch all modules
+          // that fall through the other loaders.
+          {
+            // Exclude `js` files to keep "css" loader working as it injects
+            // it's runtime that would otherwise processed through "file" loader.
+            // Also exclude `html` and `json` extensions so they get processed
+            // by webpacks internal loaders.
+            exclude: [/\.js$/, /\.html$/, /\.json$/],
+            loader: require.resolve('file-loader'),
+            options: {
+              name: 'static/media/[name].[hash:8].[ext]',
+            },
+          },
+        ],
+      },
+      // ** STOP ** Are you adding a new loader?
+      // Make sure to add the new loader(s) before the "file" loader.
+    ],
+  },
+  plugins: [
+    // Makes some environment variables available in index.html.
+    // The public URL is available as %PUBLIC_URL% in index.html, e.g.:
+    // <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
+    // In development, this will be an empty string.
+    new InterpolateHtmlPlugin(env.raw),
+    // Generates an `index.html` file with the <script> injected.
+    new HtmlWebpackPlugin({
+      inject: true,
+      favicon: paths.favicon,
+      template: paths.appHtml,
+    }),
+    // Add module names to factory functions so they appear in browser profiler.
+    new webpack.NamedModulesPlugin(),
+    // Makes some environment variables available to the JS code, for example:
+    // if (process.env.NODE_ENV === 'development') { ... }. See `./env.js`.
+    new webpack.DefinePlugin(env.stringified),
+    // This is necessary to emit hot updates (currently CSS only):
+    new webpack.HotModuleReplacementPlugin(),
+    // Watcher doesn't work well if you mistype casing in a path so we use
+    // a plugin that prints an error when you attempt to do this.
+    // See https://github.com/facebookincubator/create-react-app/issues/240
+    new CaseSensitivePathsPlugin(),
+    // If you require a missing module and then `npm install` it, you still have
+    // to restart the development server for Webpack to discover it. This plugin
+    // makes the discovery automatic so you don't have to restart.
+    // See https://github.com/facebookincubator/create-react-app/issues/186
+    new WatchMissingNodeModulesPlugin(paths.appNodeModules),
+    // Moment.js is an extremely popular library that bundles large locale files
+    // by default due to how Webpack interprets its code. This is a practical
+    // solution that requires the user to opt into importing specific locales.
+    // https://github.com/jmblog/how-to-optimize-momentjs-with-webpack
+    // You can remove this if you don't use Moment.js:
+    new webpack.IgnorePlugin(/^\.\/locale$/, /moment$/),
+  ],
+  // Some libraries import Node modules but don't use them in the browser.
+  // Tell Webpack to provide empty mocks for them so importing them works.
+  node: {
+    dgram: 'empty',
+    fs: 'empty',
+    net: 'empty',
+    tls: 'empty',
+  },
+  // Turn off performance hints during development because we don't do any
+  // splitting or minification in interest of speed. These warnings become
+  // cumbersome.
+  performance: {
+    hints: false,
+  },
 };
